// src/components/ProductManagement/types.ts
export interface Product {
  id: number;
  name: string;
  vendor: string;
  category: string;
  price: string;
  discountPrice: string;
  stock: number;
  status: string;
  createdDate: string;
  image: string;
  description: string;
  store: string;
  publisher: string;
  variations: string;
  salePrice: string;
}